Enable POST-ing payloads to show in the parser
With this change, you an call /payload/send with a POST and FormData containing "payload" field.

Should solve for: https://twitter.com/matt_kruse/status/1664245298456059904

An example of how the API can be used is in `app/payload/send`.

The way this works is a bit ugly, but I don't know of a better way. What happens is:
1. A route handler extracts the payload from the FormData
2. Converts it to JSON
3. Converts it to Base64

It then creates a HTML response that:
1. Converts from Base64 to JSON string
2. Converts JSON string to string (payload)
3. Save payload to localStorage
4. Redirects to "/"

Then parser will show the localStorage data.

This is maybe some kind of security nightmare, but the constraints that led me to this are:
- Next.js page can't read POST data.
- Headers are limited in size.
- Cookies are also therefor limited in size.