Refactor fetch patch
The previous fetch patch was not exhaustive enough. For example, it did not handle `fetch(new Request(...))` properly, causing sites like YouTube to fail.

I've rewritten some of the code to be able to handle this now, and also to abort to returning the response unmodified in case enough data to run the custom stuff is not available, meaning it should break less often.

There's also now handling of plain `header` objects vs `new Header()`.

I also changed the code that copies `fetch` to `originalFetch` to make it possible to execute the script multiple times without overwriting the original `fetch` (`originalFetch`) each time.

Fixes #143