Optimize raw data viewer
Below the trees, there's been two dropdowns to view raw JSON data as a string or using `react-json-tree`.

It turns out these are quite slow to render. To improve this, I've removed `react-json-tree` (it was not that useful) and made the string render properly format the json. I also made it so that none of this runs until you expand the tag.