Move Parser component to core as PayloadViewer
This change introduces the `viewer` folder in core, meant for basically complete apps that don't need much or anything to be ran.

The new `PayloadViewer` is meant to be ran on the website, replacing `Parser`. It contains all of the interactive functionality of the website.

I've also added stories for various default payloads.