Simplify message listener in extension App.tsx
Previously, to handle incoming messages in App.tsx, I did a bunch of copies of the entire message array to deduplicate it, sort it, and remove some html responses that shouldn't be there (other problem to fix).

This was inefficient, because it had to perform these operations on every render of App (whenever a new message came in).

Instead, we can do some filtering of each incoming message to determine wether or not it should be added to the message array. This is faster. I also wrapped the `setMessages` call with `startTransition` to make it non-blocking.