Refactor fetch-patch to read the `Content-Type: text/x-component` header
Instead of looking for `RSC: 1` in the request headers (a Next.js-specific thing), we can look for `Content-Type: text/x-component` in response headers.

This works well, and will be compatible with more frameworks down the line too.

With this change, extracting the url and request headers can be moved further down into the script.

This is good because the earlier the script can return an unmodified response to the website code, the smaller the risk is that something unexpected crashes.

Now there's just three simple if cases that each have early returns in case the conditions for RSC parsing are not fulfilled.