Remove vite dev server from script exception from build manifest.json (#198)
* Remove vite dev server from script exception from build manifest.json

The way that I've set up Vite HRM relies on ` http://localhost:6020` being part of `content_security_policy.extension_pages` when developing. This allows the extension to talk to the dev server.

But in production, all needed files are included in the dist, and localhost should not not be mentioned in the manifest.json. So I asked ChatGPT to help me write a script to remove it from the manifest.json in the dist. And tested it locally.

Since this is the only change that needs to happen between dev and prod I think this is acceptable for now, but looking into generating the manifest.json might be a good idea later.

* Test variation of sed that might work on Linux

* Test potentially universal command