Create recording mode (#195)
Added a "Start recording" button that triggers the injection of the `fetch-patch.js` script. This is necessary to reduce the risk of bugs in the fetch patching causing bugs on any site (since the content script runs everywhere.

One drawback of this is that it's now hard to see prefetched requests that happen immediately on initial page load. Will have to think about a solution to bring that back later.

Reloading the page will now reset the saved messages and force you to re-inject the content script by starting the recording again. Also worth keeping in mind is that the messages as still global, so reloading any tab will currently stop the recording.