Fix tree data parsing when null (#280)
Recently I discovered some rows that are: `5:null` for example. The parser didn't handle this, leading to a crash in the `RowTab` rendering for those rows.

I'm not to happy about the solution. It assumes that there are no other row types that start with `n`. I haven't found any yet, but this solution likely won't hold up long-term.