Flatten `NodeOther` into `Node` (#709)
I realized that `NodeOther` could be flattened into `Node`.  This has the benefit of there only being one main place where switching is happening instead of two.

To make this easier, I created `JSContainerWrapperForObjects` that wraps `JSContainer` as needed based on the `ObjectContext`.

Then I also re-ordered the components a bit to make the order match the `NodeSwitch` more closely.