Embedded RscDevtoolsPanel component
Introducing a component called `RscDevtoolsPanel` that is exported from `@rsc-parser/embedded` (meant to be published on npm). It's

It is set up quite similar to how the @rsc-parser/chrome-extension, and uses shared components from @rsc-parser/core that were introduced in #857…

I've also created a package in a new `examples/` folder that is a Next.js project showcasing the `RscDevtoolsPanel` in use.