Don't flip recording state when running triggerReadNextJsScriptTags (#1158)
Previously, when running, triggerReadNextJsScriptTags, `setIsRecording(true);` would be called. This made the UI look like it was recording, but it actually wasn't, so navigating around didn't yield any additional captured requests.

Now, `setIsRecording(true)` won't be called, and the UI will not be in a false recording state anymore. You'll have to turn on recording yourself.