Fix dev check in the Chrome extension (#1776)
* Fix dev check in the Chrome extension

Previous, the `__DEV__` checking code in production builds of the Chrome extension was always `false`, because `process.env.NODE_ENV` was statically evaluated in build.

Now it's guessing wether it's dev or prod based on the URL of the RSC_REQUEST event instead. If `localhost` or `127.0.0.1`. This may be inaccurate sometimes, but I think it will mostly work.

* Fix formatting